+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Move `PlatformMetric` to `iexec-core-library` subproject, modify it to become immutable. (#628 #629)
 - Add prometheus endpoint with custom metrics. (#632)
 - Expose version through prometheus endpoint. (#637)
+- Stop fetching completed tasks count from DB. (#638)
 
 ### Quality